requirements.txt
torch
torchvision
PILLOW
matplotlib
numpy